+++ lib/isolator/adapters.rb
@@ -3,4 +3,5 @@
 require "isolator/adapters/http"
 require "isolator/adapters/background_jobs"
 require "isolator/adapters/mailers"
+require "isolator/adapters/websockets"
 require "isolator/adapters/after_commit" if defined?(::TestAfterCommit)