+++ spec/isolator/adapters/base_spec.rb
@@ -4,7 +4,7 @@ require "spec_helper"
 
 describe "Base adapter" do
   before(:all) do
-    module ::Isolator::Danger # rubocop:disable Lint/ConstantDefinitionInBlock
+    class ::Isolator::Danger # rubocop:disable Lint/ConstantDefinitionInBlock
       def self.call(a, b, method: :+)
         a.send(method, b)
       end
@@ -50,6 +50,45 @@ describe "Base adapter" do
       expect {
         subject.call(2, 2, method: :+)
       }.not_to raise_error
+    end
+  end
+
+  describe "#ignore_target?" do
+    before(:all) do
+      class ::Isolator::Danger # rubocop:disable Lint/ConstantDefinitionInBlock
+        attr_reader :role
+
+        def initialize(role)
+          @role = role
+        end
+
+        def perform(a, b)
+          a + b
+        end
+      end
+
+      Isolator.isolate :test_instance, target: ::Isolator::Danger,
+        method_name: :perform, ignore_on: ->(obj) { obj.role == :read }
+    end
+
+    after(:all) do
+      Isolator.adapters.delete(:test_instance)
+    end
+
+    let(:role) { :write }
+
+    subject { ::Isolator::Danger.new(role) }
+
+    specify do
+      expect { subject.perform(1, 2) }.to raise_error(Isolator::UnsafeOperationError)
+    end
+
+    context "when ignored" do
+      let(:role) { :read }
+
+      specify do
+        expect { subject.perform(1, 2) }.not_to raise_error
+      end
     end
   end
 end