+++ spec/support/active_job_init.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require "active_job/railtie"
 require "active_job"
 
 ActiveJob::Base.queue_adapter = :test